# perf: remove heap allocation in parse_host