Before:

```
test long           ... bench:         302 ns/iter (+/- 42) = 142 MB/s
```

After:

```
test long           ... bench:         271 ns/iter (+/- 7) = 158 MB/s
```